requirements.txt
Flask
twilio
google-generativeai
python-dotenv
SpeechRecognition
gTTS
requests
pydub